Fix throwing of exception in an exception handler
Function setup_metrics was throwing an exception in the exception
handler of _monitor method of TaskThread:

File "/usr/local/lib/python3.6/site-packages/celery_exporter/monitor.py", line 122, in setup_metrics
    for name, labels, cnt in metric.samples:
ValueError: too many values to unpack (expected 3)

Problem was that metrics.samples iterable contains 5 items
instead of three ('timestamp' and 'exemplar' were not anticipated).

This exception thrown in an exception handler caused hanging of the
task thread after lost of the connection to a broker. Normally
exception from the lost connection is handled in an exception handler
and recovery loop start new connection. Because of the second exception
in the handler this has never happened.

After lost of a connection to a broker, other metrics from other treads
were collected without problems.